Adds the "selected" class to the menu item that is currently selected
In a Tiki menu, the selected item is marked with the ”active” class.
When the user hovers another choice, the active class moves to this
choice and there is now no chance to know which item is currently
selected. This update adds the ”selected” class to the currently
selected item. This is only changed when a new item is selected