Better positioning of menus
In my previous fix, I moved the menu 5px up in all cases. This does not
lok good for small menus that covers the element they are aligned to
for no reason. This fix makes sure that only menus that reach the
bottom of the browser window are moved up 5 pixels.